+++ Sources/iOS/OAuth2WebViewController.swift
@@ -17,15 +17,19 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //
+#if os(iOS)
 
 import UIKit
+#if !NO_MODULE_IMPORT
+import Base
+#endif
 
 
 /**
-    A simple iOS web view controller that allows you to display the login/authorization screen.
- */
-public class OAuth2WebViewController: UIViewController, UIWebViewDelegate
-{
+A simple iOS web view controller that allows you to display the login/authorization screen.
+*/
+public class OAuth2WebViewController: UIViewController, UIWebViewDelegate {
+	
 	/// Handle to the OAuth2 instance in play, only used for debug lugging at this time.
 	var oauth: OAuth2?
 	
@@ -240,3 +244,4 @@ public class OAuth2WebViewController: UIViewController, UIWebViewDelegate
 	}
 }
 
+#endif