+++ Sources/macOS/OAuth2WebViewController.swift
@@ -3,7 +3,7 @@
 //  OAuth2
 //
 //  Created by Guilherme Rambo on 18/01/16.
-//  Copyright 2014 Pascal Pfiffner
+//  Copyright 2016 Pascal Pfiffner
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -17,9 +17,13 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //
+#if os(OSX)
 
 import Cocoa
 import WebKit
+#if !NO_MODULE_IMPORT
+import Base
+#endif
 
 
 /**
@@ -33,7 +37,7 @@ public class OAuth2WebViewController: NSViewController, WKNavigationDelegate, NS
 	}
 	
 	/// Handle to the OAuth2 instance in play, only used for debug logging at this time.
-	var oauth: OAuth2?
+	var oauth: OAuth2Base?
 	
 	/// Configure the view to be shown as sheet, false by default; must be present before the view gets loaded.
 	var willBecomeSheet = false
@@ -209,7 +213,7 @@ public class OAuth2WebViewController: NSViewController, WKNavigationDelegate, NS
 		// would work as there may be URL parameters attached
 		if let url = request.url where url.scheme == interceptComponents?.scheme && url.host == interceptComponents?.host {
 			let haveComponents = URLComponents(url: url, resolvingAgainstBaseURL: true)
-			if let hp = haveComponents?.path, ip = interceptComponents?.path where hp == ip || ("/" == hp + ip) {
+			if let hp = haveComponents?.path, let ip = interceptComponents?.path where hp == ip || ("/" == hp + ip) {
 				if onIntercept!(url: url) {
 					decisionHandler(.cancel)
 				}
@@ -260,3 +264,4 @@ public class OAuth2WebViewController: NSViewController, WKNavigationDelegate, NS
 	}
 }
 
+#endif